+++ chuck-mqtt-compose/build.gradle.kts
@@ -8,7 +8,7 @@ val version = ext.get("gitVersionName")
 android {
     kotlinOptions {
         jvmTarget = JavaVersion.VERSION_11.toString()
-        freeCompilerArgs +=  "-Xjvm-default=all"
+        freeCompilerArgs += "-Xjvm-default=all"
     }
 
     buildFeatures {
@@ -17,13 +17,13 @@ android {
     }
 
     composeOptions {
-        kotlinCompilerExtensionVersion =  "1.1.0-beta03" //versions.composeVersion
+        kotlinCompilerExtensionVersion = "1.1.0-beta03" //versions.composeVersion
     }
 }
 
 ext {
     set("PUBLISH_GROUP_ID", "com.gojek.courier")
-    set("PUBLISH_ARTIFACT_ID", "chuck-mqtt-modern")
+    set("PUBLISH_ARTIFACT_ID", "chuck-mqtt-compose")
     set("PUBLISH_VERSION", ext.get("gitVersionName"))
     set("minimumCoverage", "0.0")
 }
@@ -38,42 +38,12 @@ plugins {
 }
 
 apiValidation {
-    ignoredPackages.addAll(listOf("com.gojek.chuckmqtt.internal.data.local.room", "com.gojek.chuckmqtt.internal"))
     ignoredClasses.add("com.gojek.chuckmqtt.BuildConfig")
     nonPublicMarkers.add("androidx.annotation.RestrictTo")
 }
 
 dependencies {
-    api(project(":mqtt-client"))
-    api(project(":paho"))
-    implementation(project(":courier-core-android"))
-    implementation(project(":chuck-mqtt-compose"))
-
     implementation(deps.kotlin.stdlib.core)
-
-    implementation(deps.android.gson)
-
-    implementation(deps.android.androidx.lifecycleExtensions)
-    implementation(deps.android.androidx.lifecycleCommons)
-
-    implementation(deps.android.androidx.appcompact)
-    implementation(deps.android.androidx.fragmentExtensions)
-    implementation(deps.android.androidx.coreKtx)
-    implementation(deps.android.androidx.constraintLayout)
-    implementation(deps.android.androidx.supportV4)
-    implementation(deps.android.androidx.recyclerView)
-    implementation(deps.android.androidx.material)
-
-    implementation(deps.android.room.roomRuntime)
-    implementation(deps.android.room.roomRxJava)
-    kapt(deps.android.room.roomCompiler)
-
-    implementation(deps.rx.java)
-    implementation(deps.rx.android)
-    implementation(deps.rx.rxKotlin)
-    implementation(deps.rx.rx3BindingCore)
-
-    implementation(deps.square.okio)
 
     implementation(deps.compose.composeActivity)
     implementation(deps.compose.composeAnimation)
