+++ settings.gradle
@@ -14,6 +14,7 @@ include ':courier-stream-adapter-coroutines'
 include ':adaptive-keep-alive'
 include ':network-tracker'
 include ':chuck-mqtt'
+include ':chuck-mqtt-compose'
 include ':chuck-mqtt-no-ops'
 include ':app-state-manager'
 include ':courier-auth-http'
